requirements-a2a.txt
# Google A2A (Agent-to-Agent) runner
-r requirements.txt
a2a-sdk>=0.2.0
